+++ src/locale/Locale.ts
@@ -40,6 +40,11 @@ export const SupportedLocales = Array.from(
 /** One of the supported locales above */
 export type SupportedLocale = (typeof SupportedLocales)[number];
 
+/** Placeholders in the locale template language */
+export const Unwritten = '$?';
+export const Outdated = '$!';
+export const MachineTranslated = '$~';
+
 /**
  * Represents a complete translation for Wordplay,
  * including every user interface label, every description, etc.
@@ -115,7 +120,7 @@ export type NameText = string | string[];
 export type DocText = string | string[];
 
 export function toDocString(doc: DocText) {
-    return Array.isArray(doc) ? doc.join('\n\n') : doc;
+    return withoutAnnotations(Array.isArray(doc) ? doc.join('\n\n') : doc);
 }
 
 export function parseLocaleDoc(doc: string) {
@@ -130,8 +135,24 @@ export function getFirstName(name: NameText) {
     return typeof name === 'string' ? name : name[0];
 }
 
-export function nameWithoutMentions(name: string) {
-    return name.replaceAll('$?', '').replaceAll('$!', '').trim();
+export function withoutAnnotations(name: string) {
+    return name
+        .replaceAll(Unwritten, '')
+        .replaceAll(Outdated, '')
+        .replaceAll(MachineTranslated, '')
+        .trim();
+}
+
+export function isUnwritten(text: string) {
+    return text.startsWith(Unwritten);
+}
+
+export function isOutdated(text: string) {
+    return text.startsWith(Outdated);
+}
+
+export function isAutomated(text: string) {
+    return text.startsWith(MachineTranslated);
 }
 
 export function toLocaleString(locale: Locale) {