+++ src/parser/toMarkup.ts
@@ -3,12 +3,13 @@ import type Markup from '../nodes/Markup';
 import { DOCS_SYMBOL } from './Symbols';
 import { toTokens } from './toTokens';
 import parseDoc from './parseDoc';
+import { withoutAnnotations } from '@locale/Locale';
 
 export function toMarkup(template: string): [Markup, Spaces] {
     // Replace out of date markers before parsing
     const tokens = toTokens(
         (template.startsWith(DOCS_SYMBOL) ? '' : DOCS_SYMBOL) +
-            template.replaceAll('$!', '').trim() +
+            withoutAnnotations(template) +
             (template.endsWith(DOCS_SYMBOL) ? '' : DOCS_SYMBOL),
     );
     return [parseDoc(tokens).markup, tokens.getSpaces()];